backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
aiohttp==3.9.5
redis==5.0.4
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
beautifulsoup4==4.12.3 